create _mock_utils.ts file
